chore: follow up 3433 (#3500)
* chore: remove unnecesary fallback

* chore: remove dead code

* chore: move useLimitHasEnoughAllowance to its own file
